poprawione zapisywanie klucza w encryption
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/trunk@3520 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ modules/encryption/encryption.cpp
@@ -386,7 +386,7 @@ void SavePublicKey::yesClicked()
 	QString keyfile_path;
 
 	keyfile_path.append(ggPath("keys/"));
-	keyfile_path.append(user.ID("Gadu").toUInt());
+	keyfile_path.append(user.ID("Gadu"));
 	keyfile_path.append(".pem");
 	keyfile.setName(keyfile_path);
